cloud_scale/requirements.txt
matplotlib==3.5.2
jax==0.3.15
jaxlib==0.3.15
requests==2.28.1
ipykernel==6.15.1
ipython==8.4.0
optax==0.1.3
dm-haiku==0.0.7
dm-tree==0.1.7
dm-sonnet==2.0.0
tensorflow==2.9.1
gym[box2d]==0.25.2
azure-ai-ml==0.1.0b6